+++ src/pie/bie/graph/index.ts
@@ -6,3 +6,4 @@
 export * from "./types";
 export * from "./entityResolver";
 export * from "./edgeProposalQueue";
+export * from "./graphQueryService";